requirements.txt
langchain-core>=0.3.43
langchain-groq>=0.2.5
langchain-community>=0.3.19
faiss-cpu>=1.10.0
streamlit>=1.43.1
python-dotenv>=1.0.1
youtube-transcript-api>=0.6.3
langchainhub>=0.1.21
langchain-text-splitters>=0.3.6

sentence-transformers>=2.2.2
beautifulsoup4>=4.12.2
